+++ BinanceTradingBot/BinanceTradingBot/BinanceResponses/Kline.cs
@@ -4,16 +4,16 @@ namespace BinanceTradingBot.BinanceResponses;
 public class Kline
 {
     public long OpenTime { get; set; }
-    public decimal Open { get; set; }
-    public decimal High { get; set; }
-    public decimal Low { get; set; }
-    public decimal Close { get; set; }
-    public decimal Volume { get; set; }
+    public float Open { get; set; }
+    public float High { get; set; }
+    public float Low { get; set; }
+    public float Close { get; set; }
+    public float Volume { get; set; }
     public long CloseTime { get; set; }
-    public decimal QuoteAssetVolume { get; set; }
+    public float QuoteAssetVolume { get; set; }
     public int NumberOfTrades { get; set; }
-    public decimal TakerBuyBaseAssetVolume { get; set; }
-    public decimal TakerBuyQuoteAssetVolume { get; set; }
+    public float TakerBuyBaseAssetVolume { get; set; }
+    public float TakerBuyQuoteAssetVolume { get; set; }
 }
