+++ BinanceTradingBot/BinanceTradingBot/Models/StartBotRequest.cs
@@ -5,7 +5,7 @@
         public string ApiKey { get; set; }
         public string ApiSecret { get; set; }
         public string Symbol { get; set; }
-        public decimal TradeAmount { get; set; }
+        public float TradeAmount { get; set; }
         public int Leverage { get; set; }
     }
